Refactor a bit more.
