+++ src/test/kotlin/no/nav/dagpenger/regel/api/models/InntektTest.kt
@@ -6,7 +6,9 @@ import no.nav.dagpenger.events.Packet
 import no.nav.dagpenger.events.inntekt.v1.Inntekt
 import no.nav.dagpenger.events.inntekt.v1.KlassifisertInntektMåned
 import org.junit.jupiter.api.Test
+import java.time.LocalDate
 import java.time.YearMonth
+import kotlin.test.assertEquals
 
 internal class InntektsPeriodeTest {
 
@@ -51,5 +53,26 @@ internal class InntektTest {
             it.inntektsListe shouldBe listOf()
             it.manueltRedigert shouldBe true
         }
+    }
+}
+
+internal class SenesteInntektsmånedTest {
+    @Test
+    fun `senesteInntektsMåned is two months ago when beregningsdato is before the 5th`() {
+        assertEquals(YearMonth.of(2019, 1), senesteInntektsmåned(LocalDate.of(2019, 3, 1)))
+        assertEquals(YearMonth.of(2019, 1), senesteInntektsmåned(LocalDate.of(2019, 3, 2)))
+        assertEquals(YearMonth.of(2019, 1), senesteInntektsmåned(LocalDate.of(2019, 3, 5)))
+    }
+
+    @Test
+    fun `senesteInntektsMåned is one month ago when beregningsdato is after the 5th where the 6th is a working day`() {
+        assertEquals(YearMonth.of(2019, 2), senesteInntektsmåned(LocalDate.of(2019, 3, 6)))
+        assertEquals(YearMonth.of(2019, 7), senesteInntektsmåned(LocalDate.of(2019, 8, 15)))
+    }
+
+    @Test
+    fun `senesteInntektsMåned is two months ago when beregningsdato is after the 5th and but before there has been a working day`() {
+        assertEquals(YearMonth.of(2018, 11), senesteInntektsmåned(LocalDate.of(2019, 1, 6)))
+        assertEquals(YearMonth.of(2023, 2), senesteInntektsmåned(LocalDate.of(2023, 4, 10)))
     }
 }
